feat(disclosures.views): Move peopledb.views to disclosure views
